feature[test]: refactoring issue #103
