--- mockk/common/src/test/kotlin/io/mockk/gh/Issue103Test.kt
@@ -1,30 +1,0 @@
-package io.mockk.gh
-
-import io.mockk.every
-import io.mockk.spyk
-import io.mockk.verify
-import kotlin.test.Test
-
-class Issue103Test {
-    class MyClass {
-
-        fun myPublicMethod() {
-            myPrivateMethod()
-        }
-
-        private fun myPrivateMethod() {
-        }
-    }
-
-    @Test
-    fun getAdapter() {
-        val myClass = spyk(MyClass(), recordPrivateCalls = true)
-        every { myClass invokeNoArgs "myPrivateMethod" } returns Unit
-
-        myClass.myPublicMethod()
-
-        verify {
-            myClass invokeNoArgs "myPrivateMethod"
-        }
-    }
-}
